Maximum Array Xor
Value of an array A of length L is defined as the sum of (Ai⊕i) for all 0≤i<L, where ⊕ denotes bitwise xor operation. Note that array indices start from zero.

You are given an integer N and an array A consisting of 2N integers where Ai=i for all 0≤i<2N.

Example :

For N=1, you have an array A of length 21=2 and A=[0,1].
For N=2, you have an array A of length 22=4 and A=[0,1,2,3].
You can do at most K operations on this array. In one operation, you can choose two indices i and j (0≤i,j<2N) and swap Ai and Aj (i.e. Ai becomes Aj and vice versa).

What is the maximum value of array A you can obtain after at most K operations?