Ceiling Sum
You are given two integers A,B. You have to choose an integer X in the range [minimum(A,B), maximum(A,B)] such that:

⌈B−X2⌉+⌈X−A2⌉
is maximum.

What is the maximum sum you can obtain if you choose X optimally?

Note: ⌈x⌉ : Returns the smallest integer that is greater than or equal to x (i.e rounds up to the nearest integer). For example, ⌈1.4⌉=2, ⌈5⌉=5, ⌈−1.5⌉=−1, ⌈−3⌉=−3 , ⌈0⌉=0.